requirements.txt
# Core Dependencies
# Use headless build for Streamlit Cloud (no GUI backend required)
opencv-python-headless==4.9.0.80
numpy>=1.24.0
scipy>=1.10.0
PyYAML>=6.0

# Deep Learning & Detection
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.0.0

# Tracking
deep-sort-realtime>=1.3.0
filterpy>=1.4.5

# Visualization
matplotlib>=3.8.0
Pillow>=10.0.0
plotly>=5.18.0

# Utils
scikit-learn>=1.3.0
scikit-image>=0.22.0
pandas>=2.1.0
tqdm>=4.66.0
requests>=2.31.0
python-dotenv>=1.0.0
imageio-ffmpeg==0.4.9

# Configuration
PyYAML==6.0.1

# Optional: Web Dashboard
streamlit==1.30.0
flask==3.0.0

# Optional: Metrics & Logging
tensorboard==2.15.1
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.12.1
